build: migrate entire code set to typescript
