--- middlewares/apiMiddleware.js
@@ -1,9 +1,0 @@
-exports.mwOne = (req, res, next) => {
-  req.middlewareOneExecuted = true;
-  next();
-}
-
-exports.mwTwo = (req, res, next) => {
-  req.middlewareTwoExecuted = true;
-  next();
-}
