+++ src/db.ts
@@ -1,0 +1,4 @@
+import "dotenv/config";
+import { drizzle } from "drizzle-orm/node-postgres";
+
+const db = drizzle(process.env.DATABASE_URL!);
